+++ src/main/resources/application.yml
@@ -4,10 +4,10 @@ spring:
 ---
 spring:
   datasource:
-    driver-class-name: org.h2.Driver
-    url: 'jdbc:h2:~/test'
-    username: sa
-    password:
+    url: jdbc:mysql://localhost:3306/memory-schema?useSSL=false&serverTimezone=Asia/Seoul&characterEncoding=UTF-8
+    username: memory-user
+    password: memory*1234
+    driver-class-name: com.mysql.cj.jdbc.Driver
   h2:
     console:
       enabled: true
